Add modelfiles module and move some project file functions
